clean up resolution.py and move to _io
